Create Emails.py
I put a limit on this - this is for if you decide to make your bot public and don't want people spamming too much. This could be offered as an invite reward and more invites = more emails at once etc. If this project gets noticed I'll make it so different channels get different limits and so it supports multiple emails etc. Have fun and don't annoy people with it. :)